Remove some unnecessary headers
