+++ hw/arm/prusa/stm32_common/stm32_shared.h
@@ -121,6 +121,8 @@ enum STM32_PERIPHS {
 	STM32_P_GPIO_END = STM32_P_GPIOK + 1U,
 	STM32_P_ADC_BEGIN = STM32_P_ADC1,
 	STM32_P_ADC_END = STM32_P_ADC3 + 1U,
+    STM32_P_USART_BEGIN = STM32_P_UART1,
+    STM32_P_USART_END = STM32_P_UART8 + 1U,
 };
 #undef _P
